SAK-8809
Fixed the segments reloads and erroneous checksum failures that where caused by one thread inside
tomcat causing the reload of an index while the indexer threads on the same node were performing and index
update, when lucene performs merges it will modify cfs files and hence the checksum will fail, if and 
index is reloaded when this happens the checksum will not match the stored version and so the whole index
will be reloaded from shared storage. Consequentially the index cycle will end and  the processed documents
will not appear in the index (or will not be deleted from the index.) This may also effect situations where documents
were not being deleted from the index.

The patch set also improves debugging listing the files in the segment when a checksum fails and indicating 
where the failure happened. It does not reload from the shared space any more, defering that decision to a total
failure to load the lucen index.



git-svn-id: https://source.sakaiproject.org/svn/search/trunk@23043 66ffb92e-73f9-0310-93c1-f5514f145a0a