svn merge -c 28997 https://source.sakaiproject.org/svn/search/trunk U    search-impl/impl/src/java/org/sakaiproject/search/component/dao/impl/SearchIndexBuilderWorkerDaoJdbcImpl.java U    search-impl/impl/src/java/org/sakaiproject/search/index/impl/JDBCClusterIndexStore.java U    search-api/api/src/java/org/sakaiproject/search/model/SearchBuilderItem.java svn log -r 28997 https://source.sakaiproject.org/svn/search/trunk ------------------------------------------------------------------------ r28997 | ian@caret.cam.ac.uk | 2007-04-17 09:22:49 -0400 (Tue, 17 Apr 2007) | 15 lines
SAK-9512
SAK-8420
Fixed,


When there is no index, no documents can be deleted from the index and so the state of a item
remains locked after the processDeletes phase of the index build.

The fix changes the the filter on the processAdds to ensure that PENDING_2 and LOCKED items are
considered for indexing. Without this items that were LOCKED remain locked until the index exists and then they 
get processed to PENDING_2 and finally added to the index, hence the startup problems.

There are other fixes to prevent file not found errors


------------------------------------------------------------------------


git-svn-id: https://source.sakaiproject.org/svn/search/branches/sakai_2-4-x@29180 66ffb92e-73f9-0310-93c1-f5514f145a0a