SAK-9865
This is fix for all the slow SQL performance that we have been seeing in search where there are huge numbers of items
in the queue waiting to be indexed. 

There is a new column in the searchbuilderitem table that categorizes the type of record enabling direct selection of items
site master items or the global master item, with a simple itemscope = ? in the SQL. 

The patch includes a check to see if SQK-9865.sql has been applied to the database and then migrates the old data
to the new high speed format, when the component starts up. If the pach has not been applied the search component will
exit the JVM during the init phase with an Error message to prevent the Sakai instance starting up and doing dammage to the table
and potentially acting strangely in operation.



git-svn-id: https://source.sakaiproject.org/svn/search/trunk@30373 66ffb92e-73f9-0310-93c1-f5514f145a0a