SAK-10594 Tests have passed Ok
This will not magically find more disk space on the node in question, but the central index should not now be removed.
I have not tested what happens if the local node runs out of disk space, since I would need to fill my test box up with dummy content.




git-svn-id: https://source.sakaiproject.org/svn/search/trunk@33575 66ffb92e-73f9-0310-93c1-f5514f145a0a