SAK-11245 http://jira.sakaiproject.org/jira/browse/SAK-11245
Now working in tomcat, indexing all fixed,
Configuration file for components written and moved to startup with that
Most of the commit is code formatting.
Discovered that its a real pain to work out which components have init, so they all have it now
Als Spring is not able to determin the dependency sequnce of registered events and out of band arrangements, perhapse 
constructor injection would simplify the startup sequence and avoid errors.




git-svn-id: https://source.sakaiproject.org/svn/search/trunk@35983 66ffb92e-73f9-0310-93c1-f5514f145a0a