SAK-12460
Fixed
Change the index close mechnism to reference count taking into account parent child relationships between indexreaders and 
index searchers, and to bind to the the request thread to ensure that files are no closed when active, and also are closed 
as soon as possible.

Ths has been tested with the unit test OpenFilesTest which can be  run while using lsof to verify the list of open files.
I have also tested this using a sakai instaance running in soak mode with apache benchmark hitting the search tool with batches of 6000
hits on 5 threads.

There is a related SAK on this commit (see JIRA) that fixes a ConcurrentModificationException in THreadLocalComponent



git-svn-id: https://source.sakaiproject.org/svn/search/trunk@39439 66ffb92e-73f9-0310-93c1-f5514f145a0a