SAK-12686
Fixed,

The shared segment merge transaction has been restructured so that it doesnt hold the connection open while the merge
is being performed. This removes the changes of a lock timeout consuming a database transaction. 

I have rested this using the SearchSoak test mvn -Dtest=SearchSoak clean test which runs a 4 node cluster indexing forever.
Looking at it through the jconsole JMX console, we are not getting additional connections in the pool and the number of open indexes
per node doesnt rise above 7.



git-svn-id: https://source.sakaiproject.org/svn/search/trunk@40096 66ffb92e-73f9-0310-93c1-f5514f145a0a